requirements.txt
fastapi==0.109.2
langchain==0.1.9
langchain_openai==0.0.7
pydantic==2.6.1
PyPDF2==3.0.1
python-dotenv==1.0.1
uvicorn==0.27.1
